requirements.txt
apiclient==1.0.4
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dj-database-url==2.2.0
Django==4.2.11
django-cors-headers==4.4.0
django-filter==24.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
Faker==26.0.0
fonttools==4.53.1
gunicorn==22.0.0
idna==3.6
importlib_resources==6.4.0
iniconfig==2.0.0
kiwisolver==1.4.5
lxml==5.2.1
marshmallow==3.17.0
matplotlib==3.9.1
mysqlclient==2.2.4
numpy==2.0.0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
PyJWT==1.7.1
pyparsing==3.1.2
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scipy==1.13.1
six==1.16.0
spyne==2.14.0
sqlparse==0.4.4
suds==1.1.2
tomli==2.0.1
transbank-sdk==5.0.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
waitress==3.0.0
whitenoise==6.7.0
zipp==3.19.2
